feat(header): allow "Contato" to open contact modal
This commit changes the header. I transformed the "contato" from anchor to a button, but it's still
a Text component. For that, I added transparent background and removed border for all Text
components, what didn't affect the other ones. I also added the pointer behavior globally for all
buttons. I decided that strategy because the current buttons are rounded and have very specific
style. Maybe I'll reconsider that in the future.