feat(projectpage): create project page using getStaticPaths
This commit creates each project's page, using Next's getStaticPaths resource. I used Next's Image
component to display the images, and that obligated me to create a new field on Projects'array. As
the layout asked for a url, I also added this atribute to Projects' array. The page got really few
content, so the Footer ended before the page. That's why I changed sime style behaviors of
WebsitePageWrapper to have a minHeight.